AST: ASTMangler should not depend on ASTDemangler
ASTDemangler depends on ClangImporter, etc.
